Extract PactStubber to its own file. Fix compilation issues. Add package objet
